requirements.txt
attrs==21.2.0
cattrs==1.6.0
hypothesis==6.12.0
